requirements.txt
flask
faiss-cpu
supabase
pyjwt[crypto]
python-dotenv
numpy
pypdf
sentence-transformers
